Make FootprintBanklinkingWithAuthToken to be FootprintBanklinking
